# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: pingo API
  version: 0.1.0
  description: |
    여행 공동 지도 API 계약. 정본은 이 파일이며 `최종기획안.md`의 절 번호를 각 엔드포인트 설명에 인용한다.
    변경 시 반드시 docs/CHANGELOG-api.md에 기록하고 프론트 담당자에게 통보한다 (CLAUDE.md 참고).
    security가 전역이므로 모든 엔드포인트는 401 UNAUTHORIZED를 반환할 수 있다.
    개별 엔드포인트에는 그 엔드포인트 고유의 에러만 명시한다.
servers:
  - url: https://api.pingo.example.com
    description: 배포 시 Cloud Run URL로 교체
security:
  - cookieAuth: []
...
paths:
...
  /maps/{mapId}/events:
    get:
      tags: [realtime]
      summary: |
        전체 채널 SSE. 지도의 모든 구성원이 구독한다.
        visibility=private인 핀·후보는 이 채널로 절대 내려가지 않는다 (5-5-1, 가드레일 1).
      parameters:
        - { $ref: "#/components/parameters/MapId" }
        - { $ref: "#/components/parameters/LastEventId" }
      responses:
        "200":
          description: 전체 채널 이벤트 스트림
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/PublicEvent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /maps/{mapId}/events/me:
    get:
      tags: [realtime]
      summary: |
        개인 채널 SSE. 요청자 본인만 구독한다. 본인의 비공개 AI 후보와 run 진행 상태가 여기로만 온다.
        채널 분리가 5-5-1을 if문이 아니라 구조로 강제한다.
      parameters:
        - { $ref: "#/components/parameters/MapId" }
        - { $ref: "#/components/parameters/LastEventId" }
      responses:
        "200":
          description: 개인 채널 이벤트 스트림
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/PrivateEvent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    MapId: { name: mapId, in: path, required: true, schema: { type: string } }
...
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      description: |
        브라우저 EventSource가 재연결 시 자동으로 보낸다. 이 seq 이후 이벤트만 재전송한다.
        event_log 보존기간을 넘긴 값이면 REST 재조회로 폴백한다 (docs/events.md).
      schema: { type: integer, format: int64 }
  responses:
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: 공통 에러 포맷. 코드 카탈로그는 docs/errors.md
      required: [code, message]
      properties:
        code: { type: string, example: "PIN_DUPLICATE" }
        message: { type: string }
        detail: { type: object, additionalProperties: true }
    Category:
      type: string
      enum: [음식점, 카페, 숙소, 관광지]
    PinKind:
      type: string
      enum: [일반, AI추천, 확정]
      description: 핀 종류는 이 3가지뿐 (기획안 9절). 확정이 나머지 둘을 덮어쓴다.
    Permissions:
      type: object
      description: 요청자 기준 계산된 권한. FE는 이 값만 보고 버튼을 그린다 (permissions.md)
      properties:
        can_react: { type: boolean }
        can_revert: { type: boolean }
        can_add_to_shortlist: { type: boolean }
        can_remove_from_shortlist: { type: boolean }
        can_disable: { type: boolean, description: "evidence_line 전용: 자기가 쓴 것만 true" }
        can_delete: { type: boolean }
...
    Member:
      type: object
      properties:
        user_id: { type: string }
        display_name: { type: string }
        online: { type: boolean }
...
    PriceBucket:
      type: string
      enum: [low, mid, high]
      description: "docs/constraints.md 차원 압축 결과. 원본 가격 숫자는 API로 노출하지 않는다"
    LabelConfidence:
      type: string
      enum: [known, unknown]
    Check:
      type: object
      description: 가드레일 5 "조건별 충족 체크"
      properties:
        fact_key: { type: string }
        label: { type: string }
        passed: { type: boolean }
        confidence: { $ref: "#/components/schemas/LabelConfidence" }
        needs_check: { type: boolean, description: "unknown_policy=pass+needs_check인 취향 조건이 unknown일 때 true" }
    Pin:
      type: object
      properties:
        id: { type: string }
        map_id: { type: string }
        category: { $ref: "#/components/schemas/Category" }
        kind: { $ref: "#/components/schemas/PinKind" }
        visibility: { type: string, enum: [public, private] }
        lat: { type: number }
        lng: { type: number }
        place_name: { type: string }
        created_by: { type: string, description: "핀을 찍은 구성원의 user_id (#26)" }
        created_by_display_name: { type: string, description: "핀 상세에서 '누가 찍었는지' 표시용 (#26)" }
        price_bucket:
          $ref: "#/components/schemas/PriceBucket"
        checks:
          type: array
          items: { $ref: "#/components/schemas/Check" }
          description: "게시된 AI 추천 핀도 상세에서 계속 노출한다 (가드레일 5)"
        source_run_id: { type: [string, "null"] }
        reaction_summary:
          type: object
          properties:
            like: { type: integer }
            neutral: { type: integer }
            against: { type: integer }
        permissions: { $ref: "#/components/schemas/Permissions" }
...
    Candidate:
      type: object
      properties:
        id: { type: string }
        place_name: { type: string }
        region_label: { type: string, description: "5-6-1 지역별 안배 태그" }
        rank: { type: integer }
        checks:
          type: array
          items: { $ref: "#/components/schemas/Check" }
        visibility: { type: string, enum: [private, published] }
        published_pin_id: { type: [string, "null"] }
...
    ShortlistItem:
      type: object
      properties:
        id: { type: string }
        pin: { $ref: "#/components/schemas/Pin" }
        visit_order: { type: [integer, "null"] }
        added_by: { type: string }
        permissions: { $ref: "#/components/schemas/Permissions" }
    Route:
      type: object
      properties:
        region_label: { type: string }
        ordered_pin_ids:
          type: array
          items: { type: string }
        total_distance_m: { type: number }
        legs:
          type: array
          items:
            type: object
            properties:
              from_pin_id: { type: string }
              to_pin_id: { type: string }
              distance_m: { type: number }
              approx_minutes: { type: integer }

    # ---------- SSE 이벤트 (docs/events.md가 정본) ----------
    SseEnvelope:
      type: object
      description: "모든 SSE 이벤트의 공통 봉투. id는 event_log.seq와 같고 SSE의 id: 필드로 전송된다"
      required: [id, event, data]
      properties:
        id: { type: integer, format: int64, description: "= event_log.seq" }
        event: { type: string }
    PublicEvent:
      description: 전체 채널 이벤트 (docs/events.md). event 값으로 data 타입이 갈린다
      oneOf:
        - $ref: "#/components/schemas/EvtPinCreated"
        - $ref: "#/components/schemas/EvtPinPublished"
        - $ref: "#/components/schemas/EvtPinDeleted"
        - $ref: "#/components/schemas/EvtReactionChanged"
        - $ref: "#/components/schemas/EvtShortlistChanged"
        - $ref: "#/components/schemas/EvtRouteRecalculated"
        - $ref: "#/components/schemas/EvtMemberJoined"
        - $ref: "#/components/schemas/EvtMemberPresence"
    PrivateEvent:
      description: 개인 채널 이벤트 (docs/events.md, 5-5-1)
      oneOf:
        - $ref: "#/components/schemas/EvtRunProgress"
        - $ref: "#/components/schemas/EvtRunCandidatesReady"
        - $ref: "#/components/schemas/EvtRunFailed"
    EvtPinCreated:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "pin.created" }
            data: { $ref: "#/components/schemas/Pin" }
    EvtPinPublished:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "pin.published" }
            data: { $ref: "#/components/schemas/Pin" }
    EvtPinDeleted:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "pin.deleted" }
            data:
              type: object
              required: [pin_id]
              properties:
                pin_id: { type: string }
    EvtReactionChanged:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "reaction.changed" }
            data:
              type: object
              required: [pin_id, reaction_summary]
              properties:
                pin_id: { type: string }
                reaction_summary:
                  type: object
                  properties:
                    like: { type: integer }
                    neutral: { type: integer }
                    against: { type: integer }
    EvtShortlistChanged:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "shortlist.changed" }
            data:
              type: object
              required: [item, action]
              properties:
                item: { $ref: "#/components/schemas/ShortlistItem" }
                action: { type: string, enum: [added, removed, reordered], description: "reordered는 수동 정렬 (#30)" }
    EvtRouteRecalculated:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "route.recalculated" }
            data:
              type: array
              items: { $ref: "#/components/schemas/Route" }
    EvtMemberJoined:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "member.joined" }
            data: { $ref: "#/components/schemas/Member" }
    EvtMemberPresence:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "member.presence" }
            data:
              type: object
              required: [user_id, online]
              properties:
                user_id: { type: string }
                online: { type: boolean }
    EvtRunProgress:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "run.progress" }
            data:
              type: object
              required: [run_id, step, label]
              properties:
                run_id: { type: string }
                step: { type: integer, minimum: 1, maximum: 8, description: "추천 8단계 (기획안 3절, 6절)" }
                label: { type: string }
    EvtRunCandidatesReady:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "run.candidates_ready" }
            data:
              type: object
              required: [run_id, candidates]
              properties:
                run_id: { type: string }
                candidates:
                  type: array
                  items: { $ref: "#/components/schemas/Candidate" }
    EvtRunFailed:
      allOf:
        - $ref: "#/components/schemas/SseEnvelope"
        - type: object
          required: [event, data]
          properties:
            event: { const: "run.failed" }
            data:
              type: object
              required: [run_id, error]
              properties:
                run_id: { type: string }
                error: { $ref: "#/components/schemas/Error" }